+++ src/sll/include/sll/api/date.h
@@ -63,6 +63,17 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_float_t sll_date_to_time(sll_date_t* dt);
 
 /**
  * \flags api func
+ * \name sll_api_date_get_time_zone
+ * \group date-api
+ * \desc Docs!
+ * \ret A
+ */
+__API_FUNC_DECL(date_get_time_zone);
+
+
+
+/**
+ * \flags api func
  * \name sll_api_date_merge
  * \group date-api
  * \desc Docs!