+++ src/sll/platform/linux/linux_util.c
@@ -98,7 +98,7 @@ void _init_platform(void){
 	}
 	sll_copy_data(nm,sz,_linux_platform_time_zone.nm);
 	_linux_platform_time_zone.nm[sz]=0;
-	_linux_platform_time_zone.off=-timezone;
+	_linux_platform_time_zone.off=-timezone/60;
 }