+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Base memory allocation function (`sll_allocate`, `sll_allocate_stack`, `sll_reallocate`, `sll_zero_allocate` and `sll_zero_allocate_stack`) raise critical errors when memory requests cannot be fulfilled ([#83])
 - C hashing API: `sll_md5_data_t`, `sll_sha1_data_t`, `sll_sha256_data_t`, `sll_sha512_data_t`, `SLL_INIT_MD5`, `SLL_INIT_SHA1`, `SLL_INIT_SHA256`, `SLL_INIT_SHA512`, `SLL_INIT_MD5_STRUCT`, `SLL_INIT_SHA1_STRUCT`, `SLL_INIT_SHA256_STRUCT`, `SLL_INIT_SHA512_STRUCT`, `sll_hash_md5`, `sll_hash_sha1`, `sll_hash_sha256` and `sll_hash_sha512`
 - Implemented `float.sll`, `float$get_compare_error`, `float$set_compare_error` and `SLL_SANDBOX_FLAG_DISABLE_FLOAT_COMPARE_ERROR_CHANGE`
-- Implemented `sll_time_zone_t`, `sll_utc_time_zone` and `sll_platform_time_zone`
+- Implemented `sll_time_zone_t`, `sll_utc_time_zone`, `sll_platform_time_zone` and `date$LOCAL_TIME_ZONE`
 
 ### Changed